Fix the type of 'input' in icccm/xcb_icccm.h/xcb_wm_hints_t
This bug behaves even worse, affecting all fields subsequent to input, if
someone insane compiles with -fpack-struct=1

+++ icccm/xcb_icccm.h
@@ -632,7 +632,7 @@ typedef struct {
   int32_t flags;
   /** Does this application rely on the window manager to get keyboard
       input? */
-  uint8_t input;
+  uint32_t input;
   /** See below */
   int32_t initial_state;
   /** Pixmap to be used as icon */
